+++ src/test/resources/debug-config/application.properties
@@ -27,3 +27,16 @@ auth.single.password=
 cleanup.INVALID=no
 cleanup.REJECTED=no
 cleanup.FAILED=no
+
+
+#
+# The URL of the deposit properties service.
+#
+easy-deposit-properties.url=http://localhost:20200/
+
+# Where to store deposit properties.
+#
+# FILE: use only the legacy deposit.properties file
+# SERVICE: use only the easy-deposit-properties service
+# BOTH: use both of the above
+easy-deposit-properties.mode=FILE
